+++ R/utils.r
@@ -1,0 +1,1 @@
+p2seg <- function(x) tibble::as_tibble(path_to_segment(x$vertex_))